pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "sentrysearch"
version = "0.1.0"
description = "Search dashcam footage using natural language queries"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "python-dotenv>=1.0",
    "chromadb>=0.4",
    "google-generativeai>=0.5",
    "ffmpeg-python>=0.2",
]

[project.scripts]
sentrysearch = "sentrysearch.cli:cli"

[tool.setuptools.packages.find]
include = ["sentrysearch*"]
